Made a first pass on the first few TODOs Updated DONE: What do we pass in here to the .loadAll function?
Updated DONE: What method do we call to render the index page?

Updated DONE: When we don't already have the rawData,
we need to retrieve the JSON file from the server with AJAX (which jQuery method is best for this?),
cache it in localStorage so we can skip the server call next time,
then load all the data into Article.all with the .loadAll function above,
and then render the index page.

Updated missing ";" in article.js on lines 41, 42, and 62

Updated missing ";" in articleView.js line 104